renamed drivers -> HAL and HAL -> controller
