--- lib/HAL/activate_buzzer.c
@@ -1,7 +1,0 @@
-#include <activate_buzzer.h>
-#include <tone.h>
-
-void activateBuzzer()
-{
-    tone_play_starwars_tuned();
-}
