+++ lib/HAL/adxl345.c
