+++ lib/HAL/dht11.c
