+++ lib/HAL/pc_comm.c
