+++ lib/HAL/pir.c
