+++ lib/HAL/tone.h
@@ -2,3 +2,4 @@
 void tone_init();
 void tone_play(uint16_t frequency, uint16_t duration);
 void tone_play_starwars();
+void tone_play_starwars_tuned();
