+++ lib/HAL/uart.c
