+++ lib/HAL/utils.h
@@ -1,0 +1,5 @@
+double parse_double(const char *string);
+void receiveMessage();
+int create_TCP_connection();
+int connect_to_wifi();
+int sendViaTCP(char *buffer);
