+++ lib/HAL/wifi.c
