+++ lib/controllers/proximity.h
@@ -7,5 +7,3 @@ uint16_t get_calibrated_value();
 void getCurrentLevel();
 
 void setThreshold(double threshold);
-
-int sendViaTCP(char *buffer);
